ready with fixes

+++ batch-models/src/main/scala/org/ekstep/analytics/dashboard/DataUtil.scala
@@ -513,7 +513,7 @@ object DataUtil extends Serializable {
         col("reviewStatus").alias("courseReviewStatus"),
         col("channel").alias("courseChannel"),
         col("lastPublishedOn").alias("courseLastPublishedOn"),
-        col("duration").alias("courseDuration"),
+        col("duration").cast(FloatType).alias("courseDuration"),
         col("leafNodesCount").alias("courseResourceCount"),
         col("lastStatusChangedOn").alias("lastStatusChangedOn"),
         col("courseOrgID")
@@ -707,8 +707,6 @@ object DataUtil extends Serializable {
 
     df = df
       .withColumn("competenciesJson", col("data.competencies_v3"))
-
-
 
     //      .withColumn("courseName", col("data.name"))
     //      .withColumn("courseStatus", col("data.status"))
